requirements.txt
Telethon~=1.25.2
elasticsearch~=8.4.3
python-dotenv~=0.21.0